refactor(tests): 테스트 더블을 도메인별로 분리
AS-IS
- Support/TestDoubles.swift에 Game/Network 관련 더블이 한 파일에 섞여 있었다
- Lobby 전용 더블이 Tests/Mocks 아래에 있어 재사용/발견이 어려웠다

TO-BE
- Support 하위에 도메인별로 분리: GameTestDoubles / NetworkTestDoubles / PermissionTestDoubles
- 타입명은 유지하고 파일 위치만 정리했다 (동작 변경 없음)

WHY
- SRP: 관심사별로 책임을 분리해 파일 목적을 명확히 한다
- KISS: 테스트 더블 위치 규칙을 단순화해 찾기/수정 비용을 낮춘다
- 협업: 단일 파일 집중 수정으로 인한 충돌 가능성을 줄인다
